requirements.txt
fastapi>=0.68.1
gunicorn>=20.1.0
motor>=2.5.1
pymongo>=3.12.0
python-dotenv>=0.19.0
requests>=2.26.0
schedule>=1.1.0
uvicorn>=0.15.0
